Add `clone.sh` command
Add a `clone.sh` command to assist in cloning repos.  Just
	`clone REPONAME`
to clone a repo, so long as you're in a correctly named directory.
Understands github and azure devops. The last two components of the
working directory must be the URL and the project name.  Examples:
```
/path/to/github.com/jetzerb
/path/to/dev.azure.com=someuser/some project
```

> Note the `=someuser` bit on the Azure DevOps URL.  The actual URL is
https://dev.azure.com/someuser, so we need to encode the slash somehow.
I've simply replaced it with an equal sign and the `clone.sh` command
assumes that.