Azure: add tests and manifest extraction
Test/verification of software in the Azure layer.
As with the Base layer, extract the software manifest and test results.

Remove the azure cli devops extension -- it's user-specific.  Could
install it into a common directory, but then if anyone installs other
extensions in the running container, they'd get stomped on when the
container restarts.  Probably the only way to do this is
- use a volume for /usr/local
- at container startup, install a list of extensions provided at `docker
  run` time